concept-tree-backend/flask-backend/requirements.txt
flask==2.3.0
Flask-CORS==4.0.0
mongoengine==0.29.1
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
